requirements.txt
exifread==3.0.0
opencv-python==4.7.0.72
numpy==1.23.5
Pillow==9.4.0
ImageHash==4.3.1
scipy==1.10.1
tqdm==4.66.1
scikit-image==0.20.0